% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rask.R
\name{ask_google}
\alias{ask_google}
\title{Send Google search for keyword. If no keyword provided, will search on the latest error or warning.}
\usage{
ask_google(keyword = NA, warning = FALSE)
}
\arguments{
\item{keyword}{A string or vector of strings.}

\item{warning}{A Boolean.}
}
\value{
NULL, opens a web browser window for the search
}
\description{
Send Google search for keyword. If no keyword provided, will search on the latest error or warning.
}
\examples{
ask_google("Pickle Rick")
ask_google()
}
