% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rask.R
\name{ask_stackoverflow}
\alias{ask_stackoverflow}
\title{Send search for keyword to Stackoverflow. If no keyword provided, will search on the latest error or warning.}
\usage{
ask_stackoverflow(keyword = NA, warning = FALSE)
}
\arguments{
\item{keyword}{A string or vector of strings.}

\item{warning}{A Boolean.}
}
\value{
NULL, opens a web browser window for the search
}
\description{
Send search for keyword to Stackoverflow. If no keyword provided, will search on the latest error or warning.
}
\examples{
ask_stackoverflow("R shiny")
ask_stackoverflow()
}
